Move inner-classes 'Location' and 'Place' to already existing classes.
