+++ simple-fb/src/main/java/com/sromku/simple/fb/entities/Location.java
@@ -1,11 +1,64 @@
 package com.sromku.simple.fb.entities;
 
-import java.util.Locale;
+import com.google.gson.annotations.SerializedName;
 
-public class Location extends IdName {
+public class Location {
 
-    @Override
-    public String toString() {
-        return String.format(Locale.US, "id=%s,name=%s", mId, mName);
+    private static final String STREET = "street";
+    private static final String CITY = "city";
+    private static final String STATE = "state";
+    private static final String COUNTRY = "country";
+    private static final String ZIP = "zip";
+    private static final String LATITUDE = "latitude";
+    private static final String LONGITUDE = "longitude";
+
+    @SerializedName(STREET)
+    private String mStreet;
+
+    @SerializedName(CITY)
+    private String mCity;
+
+    @SerializedName(STATE)
+    private String mState;
+
+    @SerializedName(COUNTRY)
+    private String mCountry;
+
+    @SerializedName(ZIP)
+    private Integer mZip;
+
+    @SerializedName(LATITUDE)
+    private Double mLatitude;
+
+    @SerializedName(LONGITUDE)
+    private Double mLongitude;
+
+    public String getStreet() {
+        return mStreet;
     }
+
+    public String getCity() {
+        return mCity;
+    }
+
+    public String getState() {
+        return mState;
+    }
+
+    public String getCountry() {
+        return mCountry;
+    }
+
+    public Integer getZip() {
+        return mZip;
+    }
+
+    public Double getLatitude() {
+        return mLatitude;
+    }
+
+    public Double getLongitude() {
+        return mLongitude;
+    }
+
 }
